Allow synchronous factories to replace async factories to simplify and reduce number of files
Condense dynamic registrations of loaders/extensions to single file to reduce number of files per loader/extension
