--- packages/dev/loaders/src/SPLAT/splatFileLoader.dynamic.ts
@@ -1,17 +1,0 @@
-import type { ISceneLoaderPluginFactory } from "core/Loading/sceneLoader";
-import { SPLATFileLoaderMetadata } from "./splatFileLoader.metadata";
-import { registerSceneLoaderPlugin } from "core/Loading/sceneLoader";
-
-/**
- * Registers the SPLAT async plugin factory, which dynamically imports and loads the SPLAT plugin on demand.
- */
-export function registerSPLATLoader() {
-    registerSceneLoaderPlugin({
-        ...SPLATFileLoaderMetadata,
-        createPlugin: async () => {
-            // eslint-disable-next-line @typescript-eslint/naming-convention
-            const { SPLATFileLoader } = await import("./splatFileLoader.plugin");
-            return new SPLATFileLoader();
-        },
-    } satisfies ISceneLoaderPluginFactory);
-}
