+++ packages/dev/loaders/src/glTF/2.0/Extensions/dynamic.ts
@@ -1,8 +1,13 @@
-import { registerKHR_materials_pbrSpecularGlossiness } from "./KHR_materials_pbrSpecularGlossiness.dynamic";
+/* eslint-disable @typescript-eslint/naming-convention */
+
+import { GLTFLoaderExtensionRegistry } from "../glTFLoaderExtensionRegistry";
 
 /**
- * Registers the built-in glTF 2.0 extensions
+ * Registers the built-in glTF 2.0 extension async factories, which dynamically imports and loads each glTF extension on demand (e.g. only when a glTF model uses the extension).
  */
 export function registerBuiltInGLTFExtensions() {
-    registerKHR_materials_pbrSpecularGlossiness();
+    GLTFLoaderExtensionRegistry.Register("KHR_materials_pbrSpecularGlossiness", true, async (loader) => {
+        const { KHR_materials_pbrSpecularGlossiness } = await import("./KHR_materials_pbrSpecularGlossiness");
+        return new KHR_materials_pbrSpecularGlossiness(loader);
+    });
 }
